rename drive-commons.js to lib.js
