+++ src/appmixer/google/drive/UploadFile/UploadFile.js
@@ -1,13 +1,13 @@
 'use strict';
 const mime = require('mime-types');
 const { google } = require('googleapis');
-const commons = require('../drive-commons');
+const lib = require('../lib');
 
 module.exports = {
 
     async receive(context) {
 
-        const auth = commons.getOauth2Client(context.auth);
+        const auth = lib.getOauth2Client(context.auth);
         const drive = google.drive({ version: 'v3', auth });
         let { fileId, fileName: fileNameInput, folder, replace, convertToDocument } = context.messages.in.content;
 
@@ -46,7 +46,7 @@ module.exports = {
         let response;
 
         if (replace) {
-            filename = commons.escapeSpecialCharacters(filename);
+            filename = lib.escapeSpecialCharacters(filename);
             const query = `name='${filename}' and parents in '${folder ? folderId : 'root'}' and trashed=false`;
             const { data } = await drive.files.list({
                 q: query
